requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
dotenv==0.9.9
gunicorn==22.0.0
packaging==24.2
psycopg2==2.9.9
python-dotenv==1.1.0
sqlparse==0.5.0
whitenoise==6.7.0
